wxMac: Compile FancyStatusBar, GuiUtil and HistoryText
HistoryText took a little bit of work to covnert mHistoryCurrent from an
int to a size_t.

+++ wx/src/HistoryText.h
@@ -32,7 +32,7 @@
 class HistoryTextCtrl : public wxTextCtrl
 {
 	std::vector<wxString> mHistoryItems;
-	int mHistoryCurrent;
+	size_t mHistoryCurrent;
 
 public:
 	HistoryTextCtrl(wxWindow* parent,
